Reorganized /scripts Directory
