+++ package.json
@@ -26,7 +26,7 @@
   "devDependencies": {
     "jstify": "0.9.0"
   },
-  "version": "1.3.0",
+  "version": "1.4.0",
   "main": "gulpfile.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
