+++ scripts/amortization/models/payment.js
