+++ scripts/amortization/templates/inputs.tpl
