+++ scripts/amortization/views/calculator.js
