+++ scripts/weighted-interest/templates/interest-row.tpl
