+++ scripts/weighted-interest/templates/weighted-interest.tpl
