+++ scripts/weighted-interest/views/interest-row.js
